int64

+++ pkg/reconciler/route/traffic/rollout.go
@@ -68,16 +68,16 @@ type RolloutParams struct {
 	// StartTime is the Unix timestamp by when (+/- reconcile precision)
 	// the Rollout has started.
 	// This is required to compute step time and deadline.
-	StartTime int `json:"starttime,omitempty"`
+	StartTime int64 `json:"starttime,omitempty"`
 
 	// NextStepTime is the Unix timestamp when the next
 	// rollout step should performed.
-	NextStepTime int `json:"nextStepTime,omitempty"`
+	NextStepTime int64 `json:"nextStepTime,omitempty"`
 
 	// StepDuration is a rounded up number of seconds how long it took
 	// for ingress to successfully move first 1% of traffic to the new revision.
 	// Note, that his number does not include any coldstart, etc timing.
-	StepDuration int `json:"stepDuration,omitempty"`
+	StepDuration int64 `json:"stepDuration,omitempty"`
 
 	// How much traffic to move in a single step.
 	StepSize int `json:"stepSize,omitempty"`
@@ -128,7 +128,7 @@ const durationSecs = 120.0
 // ObserveReady traverses the configs and the ones that are in rollout
 // but have not observed step time yet, will have it set, to
 // max(1, nowTS-cfg.StartTime).
-func (cur *Rollout) ObserveReady(nowTS int) {
+func (cur *Rollout) ObserveReady(nowTS int64) {
 	for i := range cur.Configurations {
 		c := &cur.Configurations[i]
 		if c.StepParams.StepDuration == 0 && c.StepParams.StartTime > 0 {
@@ -147,7 +147,7 @@ func (cur *Rollout) ObserveReady(nowTS int) {
 // Step will return cur if no previous state was available.
 // Second return value is the Unix timestamp in ns of the closest
 // rollout action to take or 0, if no rollout is currently scheduled.
-func (cur *Rollout) Step(prev *Rollout, nowTS int) (*Rollout, int) {
+func (cur *Rollout) Step(prev *Rollout, nowTS int64) (*Rollout, int64) {
 	if prev == nil || len(prev.Configurations) == 0 {
 		return cur, 0
 	}
@@ -165,7 +165,7 @@ func (cur *Rollout) Step(prev *Rollout, nowTS int) (*Rollout, int) {
 	}
 
 	var ret []ConfigurationRollout
-	returnTS := math.MaxInt64
+	returnTS := int64(math.MaxInt64)
 	for t, ccfgs := range currConfigs {
 		pcfgs, ok := prevConfigs[t]
 		// A new tag was added, so we have no previous state to roll from,
@@ -261,7 +261,7 @@ func adjustPercentage(goal int, cr *ConfigurationRollout) {
 
 // stepRevisions performs re-adjustment of percentages on the revisions
 // to rollout more traffic to the last one.
-func stepRevisions(goal *ConfigurationRollout, nowTS int) {
+func stepRevisions(goal *ConfigurationRollout, nowTS int64) {
 	// Not yet ready to adjust the steps or we're done
 	// (shouldn't really be here, but better be defensive).
 	if nowTS < goal.StepParams.NextStepTime || len(goal.Revisions) < 2 {
@@ -319,7 +319,7 @@ func stepRevisions(goal *ConfigurationRollout, nowTS int) {
 
 // stepConfig takes previous and goal configuration shapes and returns a new
 // config rollout, after computing the percetage allocations.
-func stepConfig(goal, prev *ConfigurationRollout, nowTS int) *ConfigurationRollout {
+func stepConfig(goal, prev *ConfigurationRollout, nowTS int64) *ConfigurationRollout {
 	pc := len(prev.Revisions)
 	ret := &ConfigurationRollout{
 		ConfigurationName: goal.ConfigurationName,
@@ -428,9 +428,9 @@ func (cur *ConfigurationRollout) computeProperties(nowTS, minStepSec, durationSe
 	// than slightly shorter.
 	stepDuration := durationSecs / numSteps * float64(time.Second)
 
-	cur.StepParams.StepDuration = int(stepDuration)
+	cur.StepParams.StepDuration = int64(stepDuration)
 	cur.StepParams.StepSize = int(stepSize)
-	cur.StepParams.NextStepTime = int(nowTS + stepDuration)
+	cur.StepParams.NextStepTime = int64(nowTS + stepDuration)
 }
 
 // sortRollout sorts the rollout based on tag so it's consistent
